requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
fastapi==0.115.11
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.43.1
fsspec==2025.2.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==3.0.2
matplotlib==3.8.0
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.1
opencv-python==4.11.0.86
packaging==23.2
pandas==2.1.2
parso==0.8.4
Pillow==10.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==310
pyzmq==26.3.0
requests==2.32.3
scikit-learn==1.4.2
scipy==1.11.4
setuptools==75.8.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.0
sympy==1.14.0
threadpoolctl==3.5.0
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.3.0
uvicorn==0.34.0
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
